pyproject.toml
[project]
name = "wb-parsing"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "dj-database-url>=3.0.0",
    "django>=5.2.3",
    "django-filter>=25.1",
    "djangorestframework>=3.16.0",
    "dotenv>=0.9.9",
    "drf-spectacular>=0.28.0",
    "gunicorn>=23.0.0",
    "psycopg2-binary>=2.9.10",
    "pytest>=8.4.1",
    "requests>=2.32.4",
    "tenacity>=9.1.2",
]
